\name{election_GM}
\alias{election_GM}
\docType{data}
\title{Election campaign data - Germany}
\description{Election campaign data - Germany - 2017}
\usage{data(election_GM)}
\format{
  A data frame with 1215 observations on the following 6 variables.
  \describe{
    \item{\code{country}}{a numeric vector. Country code.}
    \item{\code{year_election}}{a numeric vector. Election Year.}
    \item{\code{type_cs}}{a numeric vector. Type of core sentence.}
    \item{\code{partys_all}}{a numeric vector. Party code.}
    \item{\code{direction_add_all}}{a numeric vector. Direction.}
    \item{\code{issue_agg_all}}{a numeric vector. Issue code.}
  }
}
