\name{MDScosa-package}
\alias{MDScosa-package}
\alias{MDScosa}
\docType{package}
\title{Functions for performing MDS for election campaign data.}
\description{Functions for performing multidimensional scaling (MDS) for election campaign data (COSA data). For more data visit: www.poldem.eui.eu.}
\details{
\tabular{ll}{
Package: \tab MDScosa\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2020-04-05\cr
License: \tab GPL-2\cr
}
}
\author{
Ioana-Elena Oana [aut, cre]
Maintainer: Ioana-Elena Oana <ioana.oana@eui.eu>
}